fix(workspace): keep the message queue under the 700-line gate
Move the queue store out of the controller so the architecture
completion gate passes. Pin the remote ACP invoke surface at 19
channels after adding steer-follow-up.
