+++ src/renderer/src/stores/session-store.test.ts
@@ -5087,6 +5087,7 @@ describe('session store public contract', () => {
       'src/renderer/src/pages/workspace/workspace-conversation-items.ts',
       'src/renderer/src/pages/workspace/workspace-conversation-timeline.ts',
       'src/renderer/src/pages/workspace/workspace-message-queue-controller.ts',
+      'src/renderer/src/pages/workspace/workspace-message-queue-owner.ts',
       'src/renderer/src/pages/workspace/workspace-run-marks.ts',
       'src/renderer/src/pages/workspace/workspace-session-agent-configuration-controller.ts',
       'src/renderer/src/pages/workspace/workspace-session-controller.ts',
